Learning about Aggregation: fancy word for performing operations on multiple variables at once.
Learnt about Standard Deviation and Variance.

I understand Variance as the range between numbers. (this is likely incorrect), and Standard deviation is how spread out a group of numbers are.

Also learnt about using Numpy Aggregation methods on numpy variables because of how performant they are.

And seeing as numpy methods can be used on Python variables datatypes, I don't see any reason why someone would choose to use a python method over numpy. (let's see if we get any reason for this in the future).

Worked with an array of a Hundred Thousand, that is a very large array, and it took milleseconds to perform an operation... Interesting...