Understood how to convert categorical values into numerical values.
We used the one-hot encoder to do this.
Basically, transforms the categories into different columns, and adds 1/0 depending on if the row falss into the category.

I guess the more the category that exist in the dataset, the more the columns that's produced when the data is transformed using the one-hot encoder.

Pretty interesting way of doing it. I'm pretty sure there might be other ways of doing this.. I'm very much looking forward to seeing other ways.

And, also, saw that the model was able to estimate the Prices, even though the score was pretty low.
Not sure how the Models to use are picked.
Would love to find out.

I hope the process of doing this is not experimental, and that there are guidelines to follow when picking the estimators.

Well, looking forward to learning about it..