Have a better understanding and clarity on how to choose machine learning models depending on the problem that I'm trying to solve and the dataset available to me.
Also understood better the differences between classification problems & regression problems.

Classification, as the name implies, means trying to figure out if something is this or that. Like the heart disease model for example, we're trying to figure out, does this patient have heart disease or not.
Another example might be, is this car blue / red, I know this pretty contrived, but it's what is.

Regression, is, not quite sure, how the name fits, but it's when we' trying to estimate a quantity based on the features of the data.
Mostly used for estimating prices. Like for example the car-sales dataset, now the boston house prices.
We all used regression. And it seems these RandomForestRegressor is one of the best estimators to use for solving regression problems.

So, according to what I googled,
Regression is a way of finding which variables depend on each other, finding trends in the data, and also understanding the relationship between features in the data set.

Not sure, how this applies to estimating the prices / quantity..

I need to learn Mathematics to make sense of all this.